pyproject.toml
[tool.poetry]
name = "responder-hosting"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.15.0"
aiohttp = "^3.7.4"
aiosqlite = "^0.17.0"
discord = "^1.0.1"
jishaku = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
